Refactored code into features instead of components.
Merged John's and my user implementation
