+++ src/main/java/com/blu/core/BluApplication.java
@@ -1,7 +1,6 @@
-package com.blu.backend;
+package com.blu.core;
 
 import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
@@ -11,11 +10,9 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 //scan for all entities in all packages (maybe just set to com.blu.model?)
 @EntityScan("com.blu.*")
 //enable repository package
-@EnableJpaRepositories("com.blu.repository")
-public class BluBackendApplication {
-
+@EnableJpaRepositories("com.blu.*")
+public class BluApplication {
     public static void main(String[] args) {
-        SpringApplication.run(BluBackendApplication.class, args);
+        SpringApplication.run(BluApplication.class, args);
     }
-
 }
