+++ src/main/java/com/blu/device/RegisteredDeviceRepository.java
@@ -1,8 +1,6 @@
-package com.blu.repository;
+package com.blu.device;
 
-import com.blu.model.RegisteredDevice;
 import org.springframework.data.jpa.repository.JpaRepository;
-import com.blu.model.User;
 import org.springframework.stereotype.Repository;
 
 @Repository
