--- src/main/java/com/blu/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package com.blu.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import com.blu.model.User;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-}
