--- src/main/java/com/blu/service/UserService.java
@@ -1,38 +1,0 @@
-package com.blu.service;
-import com.blu.model.User;
-import com.blu.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.Optional;
-
-@Service
-public class UserService {
-
-    @Autowired
-    private UserRepository userRepository;
-
-    public List<User> getAllUsers() {
-        return userRepository.findAll();
-    }
-
-    public Optional<User> getUserById(Long id) {
-        return userRepository.findById(id);
-    }
-
-    public User saveUser(User user) {
-        return userRepository.save(user);
-    }
-
-    public void deleteUser(Long id) {
-        userRepository.deleteById(id);
-    }
-
-    public User updateUser(Long id, User userDetails) {
-        User user = userRepository.findById(id).orElseThrow();
-        // Insert data from the controller
-        return userRepository.save(user);
-    }
-
-}
