+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
 spring.application.name=blu-backend
-spring.datasource.url=jdbc:mariadb://localhost:3306/bludatabase
-spring.datasource.username=root
-spring.datasource.password=root
+spring.datasource.url=jdbc:mariadb://localhost:3306/blu_db
+spring.datasource.username=spring
+spring.datasource.password=springboot
 spring.datasource.driver-class-name=org.mariadb.jdbc.Driver
 spring.jpa.hibernate.ddl-auto=create-drop
