+++ src/test/java/com/blu/core/BluBackendApplicationTests.java
@@ -1,4 +1,4 @@
-package com.blu.backend;
+package com.blu.core;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
